+++ lib/common/utils/data.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+/// Only used by Dart, not the actual app name.
+///
+/// Use the https://pub.dev/packages/rename package for
+/// the app name on the build instead.
+///
+/// It will be set to `'App Name'` if not initialized beforehand
+late String appName;
+
+/// Your primary app color in case you need it.
+///
+/// It will be set to `Colors.blue` if not initialized beforehand
+late Color appPrimaryColor;
+
+/// Your secondary app color in case you need it.
+///
+/// It will be set to `null` if not initialized beforehand
+Color? appSecondaryColor;